修复逻辑错误和编译问题，为所有源文件添加注释

+++ Project1/计算圆的面积.c
@@ -1,8 +1,9 @@
+// 计算圆的面积（通过半径 r 计算 πr²）
 #define _CRT_SECURE_NO_WARNINGS  1
 #pragma warning(disable:6031)
 #define PEI 3.1415926
 #include<stdio.h>
-int yuandemianji(double r)
+double yuandemianji(double r)
 {
 	double s;
 	//printf("please enter r");
